Add files via upload
A US-based housing company has decided to enter the Australian market. The company uses data analytics to purchase houses at a price below their actual values and flip them on at a higher price. For the same purpose, the company has collected a data set from the sale of houses in Australia. 
The company is looking at prospective properties to buy to enter the market. I was required to build a regression model using regularisation in order to predict the actual value of the prospective properties and decide whether to invest in them or not.